requirements.txt
# Core dependencies
pandas>=2.0.0
requests>=2.31.0
PyYAML>=6.0
yfinance>=0.2.0

# Optional: For better date handling
python-dateutil>=2.8.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
